+++ test/index.test.js
@@ -12,14 +12,13 @@
 
 /* eslint-env mocha */
 
-import chai from 'chai';
+import { expect, use } from 'chai';
 import sinon from 'sinon';
 import sinonChai from 'sinon-chai';
 import { Request } from '@adobe/fetch';
 import { main } from '../src/index.js';
 
-chai.use(sinonChai);
-const { expect } = chai;
+use(sinonChai);
 
 const sandbox = sinon.createSandbox();
 
@@ -30,8 +29,8 @@ describe('Index Tests', () => {
 
   beforeEach('setup', () => {
     messageBodyJson = {
-      type: 'cwv',
-      url: 'adobe.com',
+      type: 'dummy',
+      url: 'site-id',
       auditContext: {
         key: 'value',
       },
@@ -81,10 +80,8 @@ describe('Index Tests', () => {
     expect(resp.status).to.equal(500);
   });
 
-  it('fails when missing required env variables', async () => {
+  it('happy path', async () => {
     const resp = await main(request, context);
-
-    expect(resp.status).to.equal(500);
-    expect(resp.headers.plain()['x-error']).to.equal('internal server error');
+    expect(resp.status).to.equal(200);
   });
 });